requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
pandas>=2.0.0
numpy>=1.20.0
geopy>=2.2.0
openpyxl>=3.0.9
python-multipart>=0.0.5
requests>=2.26.0
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=2.0.0
streamlit>=1.32.0
flask==2.0.1
geopy==2.2.0
folium==0.12.1
pandas>=2.0.0
python-dotenv==0.19.0 
werkzeug==2.0.3